MGS-5849 [#ccc] Updated ovxlib for 6.4.3.p0 relase
Update ovxlib for 6.4.3.p0 release, use the whole package of ovxlib for p0
since there is a dev branch merged into mainline in VSI side, that break our "git am".

Update to nn-1.1.7.

+++ ovxlib/src/client/vsi_nn_vxkernel.c
@@ -67,7 +67,7 @@ uint8_t * vsi_nn_LoadBinarySource
     fseek( fp, 0, SEEK_SET );
 
     buf = (uint8_t *)malloc( len + 1 );
-    n = fread( buf, 1, len, fp );
+    n = (int32_t)fread( buf, 1, len, fp );
     fclose( fp );
 
     if( n != len )
